% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_cox_boot.R
\name{CoxEP}
\alias{CoxEP}
\title{Point estimate of the exceedance probability for Cox model parameters}
\usage{
CoxEP(data, surv_formula, j, cutoff, n, m, lower_tail)
}
\arguments{
\item{data}{survival data (data.frame)}

\item{surv_formula}{Survival formula}

\item{j}{Index of parameter for which the exceedance probability is obtained}

\item{cutoff}{Cutoff values (scalar or vector)}

\item{n}{Number of observations in the original study}

\item{m}{Number of observations in the replication study (defaults to n if NULL)}

\item{lower_tail}{If TRUE, reports lower tail probabilities}
}
\value{
point (scalar or vector) Point estimate of exceedance probability
}
\description{
This function obtains point estimates for the exceedance probability for Cox model parameters.
}
