% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.R
\name{exceedProb}
\alias{exceedProb}
\title{Confidence intervals for the exceedance probability}
\usage{
exceedProb(cutoff, theta_hat, sd_hat, d, alpha, n, m, interval = c(-100,
  100))
}
\arguments{
\item{cutoff}{Cutoff values (scalar or vector)}

\item{theta_hat}{Point estimate of parameter of interest}

\item{sd_hat}{Estimated standard deviation for parameter of interest}

\item{d}{Number of parameters in general linear model}

\item{alpha}{Significance level}

\item{n}{Number of observations in initial study}

\item{m}{Number of observations in replication study}

\item{interval}{Interval within which to search for roots}
}
\value{
ep Exceedance probability with confidence intervals (vector if cutoff is scalar and matrix otherwise)
}
\description{
This function obtains confidence intervals for exceedance probability
}
\examples{
library(exceedProb)

n <- 100
x <- rnorm(n = n)
theta_hat <- mean(x)
sd_hat <- sd(x)

cutoff <- seq(from = theta_hat - 0.5, to = theta_hat + 0.5, by = 0.1)

exceedProb(cutoff = cutoff, 
           theta_hat = theta_hat, 
           sd_hat = sd_hat, 
           d = 1,
           alpha = 0.05, 
           n = n,
           m = n)
}
