% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep_boot_cox.R
\name{exceedProbCoxBoot}
\alias{exceedProbCoxBoot}
\title{Bootstrap confidence intervals for the exceedance probability of Cox model parameters}
\usage{
exceedProbCoxBoot(data, surv_formula, j, cutoff = NULL, m = NULL,
  alpha = 0.05, B = 1000, lower_tail = FALSE)
}
\arguments{
\item{data}{survival data (data.frame)}

\item{surv_formula}{Survival formula (character), e.g. "Surv(time, event) ~ group"}

\item{j}{Index of parameter for which the exceedance probability is obtained}

\item{cutoff}{Cutoff values (scalar or vector if supplied, otherwise set to +/- 0.5 of theta_hat)}

\item{m}{Number of observations in the replication study (defaults to n if NULL)}

\item{alpha}{Significance level}

\item{B}{Number of bootstrap resamples}

\item{lower_tail}{If TRUE, reports lower tail probabilities}
}
\value{
ep Exceedance probability with confidence intervals
}
\description{
This function obtains nonparametric bootstrap percentile confidence for Cox model parameters.
Beta version.
}
\examples{
library(exceedProb)

# Cox model -------------------------------------------------------

# Simulate exponential data
n <- 100
m <- 100
baseline_hazard <- 1
theta <- 0.4
p_censor <- 0.3
prop_tx <- 0.5

tx_indicator = rbinom(n = n, size = 1, prob = prop_tx)
event_rate <- baseline_hazard * exp(theta * tx_indicator)
censor_rate <- event_rate * p_censor / (1 - p_censor)

event_time <- rexp(n = n, rate = event_rate)
censor_time <- rexp(n = n, rate = censor_rate)

time <- pmin(event_time, censor_time)
event <- time == event_time

surv_data = data.frame(time = time, event = event, group = tx_indicator)

# Get exceedance probability with nonparametric bootstrap percentile confidence intervals
ep <- exceedProbCoxBoot(data = surv_data,
                        surv_formula = "Surv(time, event) ~ group",
                        j = 1,
                        B = 1000)
}
