% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.R
\name{getDeltaCI}
\alias{getDeltaCI}
\title{Confidence intervals for noncentrality parameter of t-distribution}
\usage{
getDeltaCI(test_stat, alpha, d, n, interval)
}
\arguments{
\item{test_stat}{Test statistics}

\item{alpha}{Significance level}

\item{d}{Number of parameters in general linear model}

\item{n}{Number of observations in initial study}

\item{interval}{Interval within which to search for roots}
}
\value{
ep Exceedance probability with confidence intervals (vector if cutoff is scalar and matrix otherwise)
}
\description{
This function obtains confidence intervals for the non-centrality
parameter of a t-distribution.
}
