% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ep.R
\name{tRoot}
\alias{tRoot}
\title{This function is used to find the root for a t-distribution pivotal quantity}
\usage{
tRoot(delta, test_stat, df, conf_level)
}
\arguments{
\item{delta}{Non-centrality parameter}

\item{test_stat}{Test statistic at which to evaluate the t-distribution}

\item{df}{Degrees of freedom}

\item{conf_level}{Confidence level (usually alpha/2 or 1-alpha/2)}
}
\value{
dif Difference between t-distribution quantile and confidence level
}
\description{
This function returns the difference between the lower tail
probability of a non-central t-distribution and a confidence level q
where the t-distribution has df degrees of freedom and 
non-centrality parameter delta.
}
